Move DurationDropdownButton widget to a separate file
