+++ lib/view/pages/privacy/screen_saver_page.dart
@@ -3,8 +3,8 @@ import 'package:provider/provider.dart';
 import 'package:settings/constants.dart';
 import 'package:settings/l10n/l10n.dart';
 import 'package:settings/services/settings_service.dart';
+import 'package:settings/view/duration_dropdown_button.dart';
 import 'package:settings/view/pages/power/power_settings.dart';
-import 'package:settings/view/pages/power/power_settings_widgets.dart';
 import 'package:settings/view/pages/privacy/screen_saver_model.dart';
 import 'package:settings/view/section_description.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
